homework_autograder_web/requirements.txt
flask
pytesseract
pillow
pdf2image
